+++ packages/alice/lib/helper/alice_export_helper.dart
@@ -1,155 +1,126 @@
 // ignore_for_file: use_build_context_synchronously
 
 import 'dart:convert' show JsonEncoder;
-import 'dart:io' show Directory, File, FileMode, IOSink, Platform;
+import 'dart:io' show Directory, File, FileMode, IOSink;
 
 import 'package:alice/core/alice_utils.dart';
 import 'package:alice/helper/alice_conversion_helper.dart';
 import 'package:alice/helper/operating_system.dart';
+import 'package:alice/model/alice_export_result.dart';
 import 'package:alice/model/alice_http_call.dart';
 import 'package:alice/model/alice_translation.dart';
 import 'package:alice/ui/common/alice_context_ext.dart';
-import 'package:alice/ui/common/alice_dialog.dart';
 import 'package:alice/utils/alice_parser.dart';
 import 'package:alice/utils/curl.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:open_filex/open_filex.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:permission_handler/permission_handler.dart';
+import 'package:share_plus/share_plus.dart';
 
-class AliceSaveHelper {
+class AliceExportHelper {
   static const JsonEncoder _encoder = JsonEncoder.withIndent('  ');
 
-  /// Top level method used to save calls to file
-  static void saveCalls(
-    BuildContext context,
-    List<AliceHttpCall> calls,
-  ) {
-    _checkPermissions(context, calls);
+  /// Format log based on [call] and tries to share it.
+  static Future<AliceExportResult> shareCall({
+    required BuildContext context,
+    required AliceHttpCall call,
+  }) async {
+    final callLog =
+        await AliceExportHelper.buildFullCallLog(call: call, context: context);
+
+    if (callLog == null) {
+      return AliceExportResult(
+        success: false,
+        error: AliceExportResultError.logGenerate,
+      );
+    }
+
+    await Share.share(
+      callLog,
+      subject: context.i18n(AliceTranslationKey.emailSubject),
+    );
+
+    return AliceExportResult(success: true);
   }
 
+  /// Format log based on [calls] and saves it to file.
+  static Future<AliceExportResult> saveCallsToFile(
+    BuildContext context,
+    List<AliceHttpCall> calls,
+  ) async {
+    final bool permissionStatus = await _getPermissionStatus();
+    if (!permissionStatus) {
+      final bool status = await _requestPermission();
+      if (!status) {
+        return AliceExportResult(
+          success: false,
+          error: AliceExportResultError.permission,
+        );
+      }
+    }
+
+    return await _saveToFile(context, calls);
+  }
+
+  /// Returns current storage permission status. Checks permission for iOS
+  /// For other platforms it returns true.
   static Future<bool> _getPermissionStatus() async {
-    if (Platform.isAndroid || Platform.isIOS) {
+    if (OperatingSystem.isIOS) {
       return Permission.storage.status.isGranted;
     } else {
       return true;
     }
   }
 
+  /// Requests permissions for storage for iOS. For other platforms it doesn't
+  /// make any action and returns true.
   static Future<bool> _requestPermission() async {
-    if (Platform.isAndroid || Platform.isIOS) {
+    if (OperatingSystem.isIOS) {
       return Permission.storage.request().isGranted;
     } else {
       return true;
     }
   }
 
-  static Future<void> _checkPermissions(
-    BuildContext context,
-    List<AliceHttpCall> calls,
-  ) async {
-    final bool permissionStatus = await _getPermissionStatus();
-
-    if (!context.mounted) return;
-
-    if (permissionStatus) {
-      await _saveToFile(context, calls);
-    } else {
-      final bool status = await _requestPermission();
-
-      if (!context.mounted) return;
-
-      if (status) {
-        await _saveToFile(context, calls);
-      } else {
-        AliceGeneralDialog.show(
-          context: context,
-          title:
-              context.i18n(AliceTranslationKey.saveDialogPermissionErrorTitle),
-          description: context
-              .i18n(AliceTranslationKey.saveDialogPermissionErrorDescription),
-        );
-      }
-    }
-  }
-
-  static Future<String> _saveToFile(
+  /// Saves [calls] to file. For android it uses external storage directory and
+  /// for ios it uses application documents directory.
+  static Future<AliceExportResult> _saveToFile(
     BuildContext context,
     List<AliceHttpCall> calls,
   ) async {
     try {
       if (calls.isEmpty) {
-        AliceGeneralDialog.show(
-          context: context,
-          title: context.i18n(AliceTranslationKey.saveDialogEmptyErrorTitle),
-          description:
-              context.i18n(AliceTranslationKey.saveDialogEmptyErrorDescription),
-        );
-        return '';
+        return AliceExportResult(
+            success: false, error: AliceExportResultError.empty);
       }
 
-      final Directory? externalDir = switch (Platform.operatingSystem) {
-        OperatingSystem.android => await getExternalStorageDirectory(),
-        OperatingSystem.ios => await getApplicationDocumentsDirectory(),
-        _ => await getApplicationCacheDirectory(),
-      };
-
-      if (externalDir != null) {
-        final String fileName =
-            'alice_log_${DateTime.now().millisecondsSinceEpoch}.txt';
-        final File file = File('${externalDir.path}/$fileName')..createSync();
-        final IOSink sink = file.openWrite(mode: FileMode.append)
-          ..write(await _buildAliceLog(context: context));
-        for (final AliceHttpCall call in calls) {
-          sink.write(_buildCallLog(context: context, call: call));
-        }
-        await sink.flush();
-        await sink.close();
-
-        if (context.mounted) {
-          AliceGeneralDialog.show(
-            context: context,
-            title: context.i18n(AliceTranslationKey.saveSuccessTitle),
-            description: context
-                .i18n(AliceTranslationKey.saveSuccessDescription)
-                .replaceAll("[path]", file.path),
-            secondButtonTitle: Platform.isAndroid
-                ? context.i18n(AliceTranslationKey.saveSuccessView)
-                : null,
-            secondButtonAction: () =>
-                Platform.isAndroid ? OpenFilex.open(file.path) : null,
-          );
-        }
-
-        return file.path;
-      } else {
-        if (context.mounted) {
-          AliceGeneralDialog.show(
-            context: context,
-            title:
-                context.i18n(AliceTranslationKey.saveDialogFileSaveErrorTitle),
-            description: context
-                .i18n(AliceTranslationKey.saveDialogFileSaveErrorDescription),
-          );
-        }
+      final Directory externalDir = await getApplicationCacheDirectory();
+      final String fileName =
+          'alice_log_${DateTime.now().millisecondsSinceEpoch}.txt';
+      final File file = File('${externalDir.path}/$fileName')..createSync();
+      final IOSink sink = file.openWrite(mode: FileMode.append)
+        ..write(await _buildAliceLog(context: context));
+      for (final AliceHttpCall call in calls) {
+        sink.write(_buildCallLog(context: context, call: call));
       }
+      await sink.flush();
+      await sink.close();
+
+      return AliceExportResult(
+        success: true,
+        path: file.path,
+      );
     } catch (exception) {
-      if (context.mounted) {
-        AliceGeneralDialog.show(
-          context: context,
-          title: context.i18n(AliceTranslationKey.saveDialogFileSaveErrorTitle),
-          description: context
-              .i18n(AliceTranslationKey.saveDialogFileSaveErrorDescription),
-        );
-        AliceUtils.log(exception.toString());
-      }
+      AliceUtils.log(exception.toString());
+      return AliceExportResult(
+        success: false,
+        error: AliceExportResultError.file,
+      );
     }
-
-    return '';
   }
 
+  /// Builds log string based on data collected from package info.
   static Future<String> _buildAliceLog({required BuildContext context}) async {
     final PackageInfo packageInfo = await PackageInfo.fromPlatform();
 
@@ -162,8 +133,11 @@ class AliceSaveHelper {
         '\n';
   }
 
-  static String _buildCallLog(
-      {required BuildContext context, required AliceHttpCall call}) {
+  /// Build log string based on [call].
+  static String _buildCallLog({
+    required BuildContext context,
+    required AliceHttpCall call,
+  }) {
     final StringBuffer stringBuffer = StringBuffer()
       ..writeAll([
         '===========================================\n',
@@ -196,7 +170,7 @@ class AliceSaveHelper {
 
     stringBuffer.writeAll([
       '${context.i18n(AliceTranslationKey.saveLogRequestSize)} ${AliceConversionHelper.formatBytes(call.request?.size ?? 0)}\n',
-      '${context.i18n(AliceTranslationKey.saveLogRequestBody)} ${AliceBodyParser.formatBody(context: context, body: call.request?.body, contentType: call.request?.contentType)}\n',
+      '${context.i18n(AliceTranslationKey.saveLogRequestBody)} ${AliceParser.formatBody(context: context, body: call.request?.body, contentType: call.request?.contentType)}\n',
       '--------------------------------------------\n',
       '${context.i18n(AliceTranslationKey.saveLogResponse)}\n',
       '--------------------------------------------\n',
@@ -204,7 +178,7 @@ class AliceSaveHelper {
       '${context.i18n(AliceTranslationKey.saveLogResponseStatus)} ${call.response?.status}\n',
       '${context.i18n(AliceTranslationKey.saveLogResponseSize)} ${AliceConversionHelper.formatBytes(call.response?.size ?? 0)}\n',
       '${context.i18n(AliceTranslationKey.saveLogResponseHeaders)} ${_encoder.convert(call.response?.headers)}\n',
-      '${context.i18n(AliceTranslationKey.saveLogResponseBody)} ${AliceBodyParser.formatBody(context: context, body: call.response?.body, contentType: AliceBodyParser.getContentType(context: context, headers: call.response?.headers))}\n',
+      '${context.i18n(AliceTranslationKey.saveLogResponseBody)} ${AliceParser.formatBody(context: context, body: call.response?.body, contentType: AliceParser.getContentType(context: context, headers: call.response?.headers))}\n',
     ]);
 
     if (call.error != null) {
@@ -234,8 +208,11 @@ class AliceSaveHelper {
     return stringBuffer.toString();
   }
 
-  static Future<String> buildCallLog(
-      {required BuildContext context, required AliceHttpCall call}) async {
+  /// Builds full call log string (package info log and call log).
+  static Future<String?> buildFullCallLog({
+    required BuildContext context,
+    required AliceHttpCall call,
+  }) async {
     try {
       return await _buildAliceLog(context: context) +
           _buildCallLog(
@@ -243,7 +220,8 @@ class AliceSaveHelper {
             context: context,
           );
     } catch (exception) {
-      return 'Failed to generate call log';
+      AliceUtils.log("Failed to generate call log: $exception");
+      return null;
     }
   }
 }