+++ packages/alice/lib/model/alice_http_call.dart
@@ -1,9 +1,10 @@
 import 'package:alice/model/alice_http_error.dart';
 import 'package:alice/model/alice_http_request.dart';
 import 'package:alice/model/alice_http_response.dart';
+import 'package:equatable/equatable.dart';
 
 /// Definition of http calls data holder.
-class AliceHttpCall {
+class AliceHttpCall with EquatableMixin {
   AliceHttpCall(this.id) {
     loading = true;
     createdTime = DateTime.now();
@@ -28,4 +29,21 @@ class AliceHttpCall {
     this.response = response;
     loading = false;
   }
+
+  @override
+  List<Object?> get props => [
+        id,
+        createdTime,
+        client,
+        loading,
+        secure,
+        method,
+        endpoint,
+        server,
+        uri,
+        duration,
+        request,
+        response,
+        error
+      ];
 }