+++ packages/alice/lib/model/alice_http_error.dart
@@ -1,5 +1,10 @@
+import 'package:equatable/equatable.dart';
+
 /// Definition of http error data holder.
-class AliceHttpError {
+class AliceHttpError with EquatableMixin {
   dynamic error;
   StackTrace? stackTrace;
+
+  @override
+  List<Object?> get props => [error, stackTrace];
 }