+++ packages/alice/lib/utils/alice_parser.dart
@@ -5,17 +5,17 @@ import 'package:alice/ui/common/alice_context_ext.dart';
 import 'package:flutter/material.dart';
 
 /// Body parser helper used to parsing body data.
-class AliceBodyParser {
+class AliceParser {
   static const String _jsonContentTypeSmall = 'content-type';
   static const String _jsonContentTypeBig = 'Content-Type';
   static const String _stream = 'Stream';
   static const String _applicationJson = 'application/json';
-  static const JsonEncoder encoder = JsonEncoder.withIndent('  ');
+  static const JsonEncoder _encoder = JsonEncoder.withIndent('  ');
 
   /// Tries to parse json. If it fails, it will return the json itself.
   static String _parseJson(dynamic json) {
     try {
-      return encoder.convert(json);
+      return _encoder.convert(json);
     } catch (_) {
       return json.toString();
     }
@@ -79,16 +79,30 @@ class AliceBodyParser {
 
   /// Get content type from [headers]. It looks for json and if it can't find
   /// it, it will return unknown content type.
-  static String? getContentType(
-      {required BuildContext context, Map<String, dynamic>? headers}) {
+  static String? getContentType({
+    required BuildContext context,
+    Map<String, String>? headers,
+  }) {
     if (headers != null) {
       if (headers.containsKey(_jsonContentTypeSmall)) {
-        return headers[_jsonContentTypeSmall] as String?;
+        return headers[_jsonContentTypeSmall];
       }
       if (headers.containsKey(_jsonContentTypeBig)) {
-        return headers[_jsonContentTypeBig] as String?;
+        return headers[_jsonContentTypeBig];
       }
     }
     return context.i18n(AliceTranslationKey.unknown);
+  }
+
+  static Map<String, String> parseHeaders({dynamic headers}) {
+    if (headers is Map<String, String>) {
+      return headers;
+    }
+
+    if (headers is Map<String, dynamic>) {
+      return headers.map((key, value) => MapEntry(key, value.toString()));
+    }
+
+    throw ArgumentError("Invalid headers value.");
   }
 }