+++ packages/alice/test/mock/alice_logger_mock.dart
@@ -1,0 +1,4 @@
+import 'package:alice/core/alice_logger.dart';
+import 'package:mocktail/mocktail.dart';
+
+class AliceLoggerMock extends Mock implements AliceLogger {}