+++ packages/alice/test/mock/alice_storage_mock.dart
@@ -1,0 +1,4 @@
+import 'package:alice/core/alice_storage.dart';
+import 'package:mocktail/mocktail.dart';
+
+class AliceStorageMock extends Mock implements AliceStorage {}