+++ packages/alice/test/mock/build_context_mock.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+import 'package:mocktail/mocktail.dart';
+
+class BuildContextMock extends Mock implements BuildContext {}