+++ packages/alice_objectbox/lib/model/cached_alice_http_call.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: must_be_immutable
+
 import 'package:alice/model/alice_http_call.dart';
 import 'package:alice/model/alice_http_error.dart';
 import 'package:alice/model/alice_http_request.dart';
@@ -110,4 +112,24 @@ class CachedAliceHttpCall implements AliceHttpCall {
     this.response = response;
     loading = false;
   }
+
+  @override
+  List<Object?> get props => [
+        id,
+        createdTime,
+        client,
+        loading,
+        secure,
+        method,
+        endpoint,
+        server,
+        uri,
+        duration,
+        request,
+        response,
+        error
+      ];
+
+  @override
+  bool? get stringify => true;
 }